requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==5.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
flake8==6.1.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.6
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mccabe==0.7.0
pep8-naming==0.13.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
zipp==3.17.0
